ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
psycopg2-binary==2.9.9
scipy==1.11.4
nltk==3.8.1
python-dotenv==1.0.0
joblib==1.3.2



